requirements.txt
# requirements.txt
# put your stuff here
opencv-python
numpy
av
flask
RPi.GPIO
Pillow
luma.lcd
spidev
